LoadTester/src/main/java/com/example/demo/LoadTestService.java
package com.example.demo;

import java.util.List;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class LoadTestService {

    @Autowired
    private LoadTestConfiguration loadTestConfiguration;

    public List<String> getUrlsFromFile(String file, String defaultUrl) {
		return null;
        // Lógica para ler URLs de um arquivo ou usar uma URL padrão
        // Retorne a lista de URLs
    }

    public LoadTestResult runLoadTest(List<String> urls, int requests, int concurrent) {
		return null;
        // Lógica para realizar os testes de carga
        // Retorne um objeto LoadTestResult com os resultados
    }

    private int makeHttpRequest(String url) {
		return 0;
        // Lógica para fazer a solicitação HTTP
        // Retorne o código de resposta
    }

	public LoadTestConfiguration getLoadTestConfiguration() {
		return loadTestConfiguration;
	}

	public void setLoadTestConfiguration(LoadTestConfiguration loadTestConfiguration) {
		this.loadTestConfiguration = loadTestConfiguration;
	}
}
